requirements.txt
bcrypt==5.0.0
blinker==1.9.0
charset-normalizer==3.4.3
click==8.3.0
et_xmlfile==2.0.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.4
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
